requirements.txt
# Core dependencies for Medical Clinic Chatbot (Educational)

# Google Gemini API
google-generativeai>=0.3.0

# FastAPI for REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0  # For database migrations

# Environment variables
python-dotenv>=1.0.0

# Date/time utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI

# Code quality
black>=23.0.0  # Code formatter
ruff>=0.1.0    # Fast linter

# Image analysis (Task: Clinical Image Analysis)
langgraph>=0.0.20        # Multi-agent orchestration
langchain>=0.1.0         # Agent framework
langchain-core>=1.0.0    # Core abstractions
huggingface-hub>=0.20.0  # HF Inference API (FREE)
Pillow>=10.0.0           # Image processing

# Transformers for MedGemma (local inference, no API quota)
transformers>=4.36.0     # Hugging Face Transformers
torch>=2.1.0             # PyTorch for model inference
accelerate>=0.25.0       # Fast model loading
sentencepiece>=0.1.99    # Tokenizer for Gemma models
